Splitted DAO
Split DAO
Added "factory method" for connecting to different databases
Implemented "singleton" via DataSource
