+++ spring-mvc-app1/src/main/java/john/project/dao/OrderListDAO.java
@@ -1,0 +1,15 @@
+package john.project.dao;
+
+import java.util.List;
+
+import john.project.models.Order;
+import john.project.models.OrderList;
+
+public interface OrderListDAO {
+    void addOrderList(Order order, int clientId);
+    void updateOrderList(Order order, int clientID);
+    void deleteOrderList(int orderID, int clientID);
+    OrderList getOrderList(int id);
+    List<OrderList> getAllOrderLists();
+
+}
