+++ spring-mvc-app1/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+db.type=postgresql
